+++ src/components/ui/carousel/carousel-utils.tsx
